added game manager

+++ Planet Game/Assets/Scripts/ProjectileController.cs
@@ -20,7 +20,7 @@ public class ProjectileController : MonoBehaviour
 
     private void OnCollisionEnter2D(Collision2D collision){
         
-        double temp_score = GameManager.Instance.getScore();
+        float temp_score = GameManager.Instance.getScore();
         temp_score+=100;
         GameManager.Instance.setScore(temp_score);
         if(collision.gameObject.tag == "Asteroid"){
